+++ DisposeGenerator/Embedded/IncludeDisposeAttribute.cs
@@ -6,6 +6,6 @@ namespace DisposeGenerator
     /// An attribute for specifying that a field should be disposed automatically.
     /// </summary>
     /// <seealso cref="Attribute"/>
-    [AttributeUsage(AttributeTargets.Field)]
+    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Property)]
     internal class IncludeDisposeAttribute : Attribute { }
 }