+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # DisposeGenerator History
 
+## 0.2.0 (16 November 2021)
+- Add support to explicitly dispose properties with `IncludeDisposeAttribute`
+- Move all attributes from the `DisposeGenerator.Attributes` to the `DisposeGenerator` namespace
+
+---
+
 ## 0.1.3 (15 November 2021)
 - Downgrade Microsoft.CodeAnalysis.CSharp to 3.9.0 to support Visual Studio 2019 16.9 and up